requirements.txt
pillow~=8.2.0
pywin32~=300
opencv-python~=4.5.1.48
numpy~=1.20.2
albumentations~=0.5.2
torch~=1.8.1
torchvision~=0.9.1